requirements.txt
PyQt5>=5.15.0
PyMuPDF>=1.20.0
Pillow>=8.0.0
pandas>=1.3.0
openpyxl>=3.0.0
openai>=1.0.0
python-dotenv>=0.19.0
camelot-py[cv]>=0.10.1
opencv-python>=4.5.0
pytesseract>=0.3.10
numpy>=1.21.0
